+++ operator-common/src/main/java/io/strimzi/operator/common/featuregates/FeatureGates.java
@@ -2,11 +2,13 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.operator.cluster;
+package io.strimzi.operator.common.featuregates;
 
 import io.strimzi.operator.common.InvalidConfigurationException;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static java.util.Arrays.asList;
 
@@ -18,7 +20,7 @@ public class FeatureGates {
 
     private static final String CONTINUE_ON_MANUAL_RU_FAILURE = "ContinueReconciliationOnManualRollingUpdateFailure";
 
-    // When adding new feature gates, do not forget to add them to allFeatureGates() and toString() methods
+    // When adding new feature gates, do not forget to add them to allFeatureGates(), toString(), equals(), and `hashCode() methods
     private final FeatureGate continueOnManualRUFailure =
         new FeatureGate(CONTINUE_ON_MANUAL_RU_FAILURE, false);
 
@@ -102,6 +104,43 @@ public class FeatureGates {
     }
 
     /**
+     * @return  Generates the value for the environment variable that can be passed to the other operators to configure
+     *          the feature gates exactly as they are set in this instance.
+     */
+    public String toEnvironmentVariable()   {
+        List<String> gateSettings = new ArrayList<>();
+
+        for (FeatureGate gate : allFeatureGates())  {
+            if (gate.isEnabledByDefault() && !gate.isEnabled()) {
+                // It is enabled by default but it is disabled now => we need to disable it
+                gateSettings.add("-" + gate.getName());
+            } else if (!gate.isEnabledByDefault() && gate.isEnabled()) {
+                // It is disabled by default but it is enabled now => we need to disable it
+                gateSettings.add("+" + gate.getName());
+            }
+        }
+
+        return String.join(",", gateSettings);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o)  {
+            return true;
+        } else if (o == null || getClass() != o.getClass()) {
+            return false;
+        } else {
+            FeatureGates other = (FeatureGates) o;
+            return Objects.equals(continueOnManualRUFailure, other.continueOnManualRUFailure);
+        }
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(continueOnManualRUFailure);
+    }
+
+    /**
      * Feature gate class represents individual feature fate
      */
     static class FeatureGate {
@@ -155,6 +194,23 @@ public class FeatureGates {
          */
         public boolean isEnabledByDefault() {
             return defaultValue;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o)  {
+                return true;
+            } else if (o == null || getClass() != o.getClass()) {
+                return false;
+            } else {
+                FeatureGate other = (FeatureGate) o;
+                return defaultValue == other.defaultValue && Objects.equals(name, other.name) && Objects.equals(value, other.value);
+            }
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(name, defaultValue, value);
         }
     }
 }