+++ topic-operator/src/test/java/io/strimzi/operator/topic/TopicControllerIT.java
@@ -15,6 +15,7 @@ import io.strimzi.api.kafka.model.common.Condition;
 import io.strimzi.api.kafka.model.topic.KafkaTopic;
 import io.strimzi.api.kafka.model.topic.KafkaTopicBuilder;
 import io.strimzi.api.kafka.model.topic.KafkaTopicStatus;
+import io.strimzi.operator.common.featuregates.FeatureGates;
 import io.strimzi.operator.common.model.Labels;
 import org.apache.kafka.clients.admin.Admin;
 import org.apache.kafka.clients.admin.AdminClientConfig;
@@ -1233,7 +1234,7 @@ class TopicControllerIT {
                 false, "", "", "", "", "",
                 false, "", "", "", "", "",
                 useFinalizer,
-                100, 100, 10, false,
+                100, 100, 10, false, new FeatureGates(""),
                 false, false, "", 9090, false, false, "", "", "",
                 "all", false);
     }
@@ -1987,7 +1988,7 @@ class TopicControllerIT {
                 false, "", "", "", "", "",
                 false, "", "", "", "", "",
                 true,
-                1, 100, 5_0000, false,
+                1, 100, 5_0000, false, new FeatureGates(""),
                 false, false, "", 9090, false, false, "", "", "",
                 "all", false);